Move template UI into actions
