+++ template/AGENTS.md
@@ -20,17 +20,17 @@ Refer to ./.agents/skills/remix/SKILL.md
 ## Starter Layout
 
 - `app/actions/controller.tsx` owns the top-level route actions
+- `app/actions/home-page.tsx` and `app/actions/document.tsx` render the route-owned starter UI
+- `app/actions/public/` contains the browser runtime entry and interactive prompt button
 - `app/routes.ts` defines the shared route contract used by server and browser modules for type-safe hrefs
 - `app/router.ts` wires routes to route handlers
 - `app/middleware/render.tsx` installs the request-scoped renderer used by actions
-- `app/ui/` holds the shared document shell and home page UI
 - `app/assets.ts` owns the server-side asset pipeline used by the asset route and renderer
-- `app/public/entry.ts` starts the browser runtime
 - Root `public/` contains static files served unchanged from the app root
 
 ## Browser And Static Files
 
-- Put browser-reachable source in a `public/` directory inside `app/`, beside its narrowest owner, such as `app/ui/public/` or `app/actions/cart/public/`.
+- Put browser-reachable source in a `public/` directory inside `app/`, beside its narrowest owner, such as `app/actions/public/` or `app/actions/cart/public/`.
 - Keep every local dependency of a browser module in a `public/` directory inside `app/`. The shared `app/routes.ts` contract is browser-readable so modules can build type-safe links with `routes.*.href(...)`; packages allowed by `app/assets.ts` are the other exception.
 - Put images, fonts, and other files that do not need compilation in the root `public/` directory. They are served from the root URL, so `public/favicon.svg` is available at `/favicon.svg`.
