+++ template/README.md
@@ -5,17 +5,17 @@ A minimal Remix application starter with a home page.
 ## Starter Shape
 
 - `app/actions/controller.tsx` owns the top-level route actions.
+- `app/actions/home-page.tsx` and `app/actions/document.tsx` render the route-owned starter UI.
+- `app/actions/public/` contains the browser runtime entry and interactive prompt button.
 - `app/routes.ts` defines the shared route contract used by server and browser modules for type-safe hrefs.
 - `app/router.ts` wires routes to handlers.
 - `app/middleware/render.tsx` installs the request-scoped renderer used by actions.
-- `app/ui/` holds the shared document shell and home page UI.
 - `app/assets.ts` owns the server-side asset pipeline used by the asset route and renderer.
-- `app/public/entry.ts` starts the browser runtime.
 - Root `public/` contains static files served unchanged from the app root.
 
 ## Browser And Static Files
 
-Browser-reachable source lives in `public/` directories inside `app/`, beside the code that owns it. For example, the starter's interactive prompt button lives in `app/ui/public/`. The asset server compiles files in these directories and requires their local dependencies to follow the same convention. The shared `app/routes.ts` contract is also browser-readable so these modules can build type-safe links with `routes.*.href(...)`; allowed packages are the other exception.
+Browser-reachable source lives in `public/` directories inside `app/`, beside the code that owns it. For example, the starter's browser entry and interactive prompt button live in `app/actions/public/`. The asset server compiles files in these directories and requires their local dependencies to follow the same convention. The shared `app/routes.ts` contract is also browser-readable so these modules can build type-safe links with `routes.*.href(...)`; allowed packages are the other exception.
 
 Static images, fonts, and other files that do not need compilation belong in the root `public/` directory. That directory is mounted at the root URL, so `public/favicon.svg` is served as `/favicon.svg`.
