+++ template/app/assets.ts
@@ -27,4 +27,4 @@ export const assetServer = createAssetServer({
   scripts: { loaders: isHmr ? [uiHmr()] : undefined },
 })
 
-export const entryHref = await assetServer.getHref('app/public/entry.ts')
+export const entryHref = await assetServer.getHref('app/actions/public/entry.ts')
